State: Move isActivatingJetpackJumpstart selector to global selector directory
