+++ client/state/jetpack/jumpstart/selectors.js
@@ -4,18 +4,6 @@
 import { get } from 'lodash';
 
 /**
- * Returns true if we are currently making a request to activate Jumpstart. False otherwise
- * Returns null if the site is unknown, or there is no information yet.
- *
- * @param  {Object}  state       Global state tree
- * @param  {Number}  siteId      The ID of the site we're querying
- * @return {?Boolean}            Whether Jumpstart is currently being activated
- */
-export function isActivatingJumpstart( state, siteId ) {
-	return get( state.jetpack.jetpackJumpstart.requests, [ siteId, 'activating' ], null );
-}
-
-/**
  * Returns true if we are currently making a request to deactivate Jumpstart. False otherwise
  * Returns null if the site is unknown, or there is no information yet.
  *
