+++ client/state/jetpack/jumpstart/test/selectors.js
@@ -7,7 +7,6 @@ import { expect } from 'chai';
  * Internal dependencies
  */
 import {
-	isActivatingJumpstart,
 	isDeactivatingJumpstart,
 	isRequestingJumpstartStatus,
 	getJumpstartStatus
@@ -18,47 +17,6 @@ import {
 } from './fixture';
 
 describe( 'selectors', () => {
-	describe( '#isActivatingJumpstart', () => {
-		it( 'should return true if jumpstart is currently being activated', () => {
-			const stateIn = {
-					jetpack: {
-						jetpackJumpstart: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 12345678;
-			const output = isActivatingJumpstart( stateIn, siteId );
-			expect( output ).to.be.true;
-		} );
-
-		it( 'should return false if jumpstart is currently not being activated', () => {
-			const stateIn = {
-					jetpack: {
-						jetpackJumpstart: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 87654321;
-			const output = isActivatingJumpstart( stateIn, siteId );
-			expect( output ).to.be.false;
-		} );
-
-		it( 'should return null if that site is not known yet', () => {
-			const stateIn = {
-					jetpack: {
-						jetpackJumpstart: {
-							requests: REQUESTS_FIXTURE
-						}
-					}
-				},
-				siteId = 88888888;
-			const output = isActivatingJumpstart( stateIn, siteId );
-			expect( output ).to.be.null;
-		} );
-	} );
-
 	describe( '#isDeactivatingJumpstart', () => {
 		it( 'should return true if jumpstart is currently being deactivated', () => {
 			const stateIn = {
