+++ client/state/selectors/test/fixtures/jetpack-jumpstart.js
@@ -1,0 +1,22 @@
+export const items = {
+	12345678: 'jumpstart_activated',
+	87654321: 'jumpstart_dismissed',
+};
+
+export const requests = {
+	12345678: {
+		activating: true,
+		deactivating: false,
+		requesting: false
+	},
+	87654321: {
+		activating: false,
+		deactivating: true,
+		requesting: false
+	},
+	11223344: {
+		activating: false,
+		deactivating: false,
+		requesting: true
+	},
+};
